# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "api/categories": {
            "get": {
                "tags": [
                    "Fetch Category"
                ],
                "summary": "list of Categories",
                "description": "Returns list of categories",
                "operationId": "Category-index",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "api/sub-categories": {
            "get": {
                "tags": [
                    "Fetch Sub-Category"
                ],
                "summary": "list of Sub-Categories",
                "description": "Returns list of sub-categories",
                "operationId": "Sub-Category-index",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/add/sub-category": {
            "post": {
                "tags": [
                    "Add Sub-Category"
                ],
                "summary": "Sub-Category Add",
                "description": "Adding sub-category according to category",
                "operationId": "Store Sub-Category",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "sub_category_name",
                                    "category_id"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "text"
                                    },
                                    "sub_category_name": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sub-Category Inserted Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Sub-Category Inserted Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Passport",
                "scheme": "bearer"
            }
        }
    }
}